+++ index.js
@@ -114,6 +114,13 @@ exports.filter = function(obj, fn, context) {
 }
 
 /**
+ * @static
+ * @method forEach
+ * @alias each
+ */
+exports.forEach = exports.each
+
+/**
  * Checks whether the given object has the given property, inherited or not.  
  * Given a set, but `undefined` property will still return `true`.
  *